移除结点类型；所有结点参与淘汰；

+++ memory/priority/priority.go
@@ -17,5 +17,5 @@ package priority
 // Priority 如果传进来的元素没有实现该接口，则默认优先级为0
 type Priority interface {
 	// GetPriority 获取元素的优先级
-	GetPriority() int64
+	GetPriority() int
 }
